Drive: Show UserError for copy/move item to Trash
