+++ src/common/api/common/error/MoveToTrashError.ts
@@ -1,0 +1,7 @@
+import { TutanotaError } from "@tutao/tutanota-error"
+
+export class MoveToTrashError extends TutanotaError {
+	constructor(message: string) {
+		super("MoveToTrashError", message)
+	}
+}
