+++ src/common/api/common/utils/ErrorUtils.ts
@@ -60,6 +60,7 @@ import { ServerModelsUnavailableError } from "../error/ServerModelsUnavailableEr
 import { AppLockAuthenticationError } from "../error/AppLockAuthenticationError"
 import { InvalidModelError } from "../error/InvalidModelError"
 import { MoveCycleError } from "../error/MoveCycleError"
+import { MoveToTrashError } from "../error/MoveToTrashError"
 
 function isErrorObjectEmpty(obj: Record<string, unknown>): boolean {
 	return Object.keys(obj).length === 0
@@ -151,6 +152,7 @@ const ErrorNameToType = {
 	MailImportError,
 	KeyVerificationMismatchError,
 	MoveCycleError,
+	MoveToTrashError,
 	Error,
 	"java.net.SocketTimeoutException": ConnectionError,
 	"java.net.SocketException": ConnectionError,
