+++ src/common/api/worker/facades/lazy/DriveFacade.ts
@@ -37,6 +37,7 @@ import { TransferId } from "../../../common/drive/DriveTypes"
 import { ProgrammingError } from "../../../common/error/ProgrammingError"
 import { NotFoundError } from "../../../common/error/RestError"
 import { MoveCycleError } from "../../../common/error/MoveCycleError"
+import { MoveToTrashError } from "../../../common/error/MoveToTrashError"
 
 export interface BreadcrumbEntry {
 	folderName: string
@@ -60,6 +61,12 @@ function isDriveFile(source: DriveFile | DriveFolder): source is DriveFile {
 export interface DriveRootFolders {
 	root: IdTuple
 	trash: IdTuple
+}
+
+export const enum DriveFolderType {
+	Regular = "0",
+	Root = "1",
+	Trash = "2",
 }
 
 /**
@@ -253,7 +260,13 @@ export class DriveFacade {
 		return this.entityClient.load(DriveFolderTypeRef, response.folder)
 	}
 
+	/**
+	 * @throws MoveToTrashError
+	 */
 	public async copyItems(files: readonly DriveFile[], folders: readonly DriveFolder[], destination: DriveFolder, renamedFiles: Map<Id, string>): Promise<Id> {
+		if (destination.type === DriveFolderType.Trash) {
+			throw new MoveToTrashError("Cannot copy to trash")
+		}
 		const fileItems = await promiseMap(files, async (file) => {
 			const sk = assertNotNull(await this.cryptoFacade.resolveSessionKey(file))
 
@@ -285,8 +298,13 @@ export class DriveFacade {
 
 	/**
 	 * @throws MoveCycleError
+	 * @throws MoveToTrashError
 	 */
 	public async move(files: readonly DriveFile[], folders: readonly DriveFolder[], destinationId: IdTuple, renamedFiles: Map<Id, string>) {
+		const destination = await this.entityClient.load(DriveFolderTypeRef, destinationId)
+		if (destination.type === DriveFolderType.Trash) {
+			throw new MoveToTrashError("Cannot move to the trash")
+		}
 		const parents = new Set((await this.getFolderParents(destinationId)).map(getElementId))
 		if (folders.some((f) => parents.has(getElementId(f)) || isSameId(f._id, destinationId))) {
 			throw new MoveCycleError(`Cannot move folder into its child ${destinationId.join("/")}`)
