+++ src/drive-app/drive/view/DriveFolderView.ts
@@ -1,5 +1,5 @@
 import m, { Children, Component, Vnode } from "mithril"
-import { DriveClipboard, DriveFolderType, SortColumn, SortingPreference } from "./DriveViewModel"
+import { DriveClipboard, SortColumn, SortingPreference } from "./DriveViewModel"
 import { DriveFolderNav, DriveSelectedItemsActions } from "./DriveFolderNav"
 import { DriveFolderContent, DriveFolderContentAttrs, DriveFolderSelectionEvents, SelectionState } from "./DriveFolderContent"
 import { DriveFolder } from "../../../common/api/entities/drive/TypeRefs"
@@ -18,6 +18,7 @@ import { modal } from "../../../common/gui/base/Modal"
 import { DropType } from "../../../common/gui/base/GuiUtils"
 import { FileActions } from "./DriveFolderContentEntry"
 import { FolderFolderItem, FolderItem, FolderItemId } from "./DriveUtils"
+import { DriveFolderType } from "../../../common/api/worker/facades/lazy/DriveFacade"
 
 export interface DriveFolderViewAttrs {
 	selection: SelectionState
