+++ src/drive-app/drive/view/DriveGuiUtils.ts
@@ -2,8 +2,8 @@ import { DropdownButtonAttrs, DropdownChildAttrs } from "../../../common/gui/bas
 import { lang, Translation } from "../../../common/misc/LanguageViewModel"
 import { Dialog } from "../../../common/gui/base/Dialog"
 import { showStandardsFileChooser } from "../../../common/file/FileController"
-import { DriveFolderType } from "./DriveViewModel"
 import { DriveFolder } from "../../../common/api/entities/drive/TypeRefs"
+import { DriveFolderType } from "../../../common/api/worker/facades/lazy/DriveFacade"
 import { FileFolderItem, FolderFolderItem, FolderItem, FolderItemId } from "./DriveUtils"
 import { DropType } from "../../../common/gui/base/GuiUtils"
 import { Icons } from "../../../common/gui/base/icons/Icons"
