+++ src/mail-app/translations/de.ts
@@ -2253,6 +2253,8 @@ export default {
 		"yourMessage_label": "Deine Nachricht",
 		"zoomIn_action": "Hereinzoomen",
 		"zoomOut_action": "Herauszoomen",
+		"cannotMoveToTrash_msg" : "Objekte können nicht in den Papierkorb verschoben werden",
+		"cannotCopyToTrash_msg": "Objekte können nicht in den Papierkorb kopiert werden",
 		"folderContent_label": "Inhalt",
 		"createNewFolderIn_label": "Neuen Ordner erstellen in {folderName}",
 		"driveNotEmpty_msg": "Dein Drive enthält noch Dateien oder Ordner. Bitte entferne diese zuerst.",
