+++ src/mail-app/translations/de_sie.ts
@@ -2253,6 +2253,8 @@ export default {
 		"yourMessage_label": "Ihre Nachricht",
 		"zoomIn_action": "Hereinzoomen",
 		"zoomOut_action": "Herauszoomen",
+		"cannotMoveToTrash_msg" : "Objekte können nicht in den Papierkorb verschoben werden",
+		"cannotCopyToTrash_msg": "Objekte können nicht in den Papierkorb kopiert werden",
 		"folderContent_label": "Inhalt",
 		"createNewFolderIn_label": "Neuen Ordner erstellen in {folderName}",
 		"driveNotEmpty_msg": "Ihr Drive enthält noch Dateien oder Ordner. Bitte entfernen Sie diese zuerst.",
