+++ test/tests/drive/DriveViewModelTest.ts
@@ -1,12 +1,12 @@
 import o from "@tutao/otest"
 import { EntityClient } from "../../../src/common/api/common/EntityClient"
-import { DriveFacade, DriveRootFolders } from "../../../src/common/api/worker/facades/lazy/DriveFacade"
+import { DriveFacade, DriveFolderType, DriveRootFolders } from "../../../src/common/api/worker/facades/lazy/DriveFacade"
 import { Router } from "../../../src/common/gui/ScopedRouter"
 import { TransferProgressDispatcher } from "../../../src/common/api/main/TransferProgressDispatcher"
 import { EventController } from "../../../src/common/api/main/EventController"
 import { LoginController } from "../../../src/common/api/main/LoginController"
 import { UserManagementFacade } from "../../../src/common/api/worker/facades/lazy/UserManagementFacade"
-import { DriveFolderType, DriveViewModel } from "../../../src/drive-app/drive/view/DriveViewModel"
+import { DriveViewModel } from "../../../src/drive-app/drive/view/DriveViewModel"
 import { matchers, object, when } from "testdouble"
 import { EntityRestClientMock } from "../api/worker/rest/EntityRestClientMock"
 import { clientInitializedTypeModelResolver, createTestEntity } from "../TestUtils"
